記帳機器人/requirements.txt
flask
gunicorn
line-bot-sdk>=3.0.0
matplotlib
psycopg2
openai